renamed styledcomps to styles
