+++ src/styles/blogPreviewStyles.js
