+++ src/styles/blogsStyles.js
