+++ src/styles/footerStyles.js
