+++ src/styles/homeStyles.js
